refactor: Suspense를 사용하기 위해 AsyncBoundary 추가 및 파일 분리
- index.lazy.tsx로 리더보드 페이지 컴포넌트 분리
